pyproject.toml
[tool.poetry]
name = "online exam management system"
version = "0.1.0"
description = ""
authors = ["Takib Yeasar <mdtakibyeasar@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.121.1"}
uvicorn = "^0.38.0"
sqlalchemy = "^2.0.44"
psycopg2-binary = "^2.9.11"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.20"
python-dotenv = "^1.2.1"
asyncpg = "^0.30.0"
alembic = "^1.17.1"
psycopg2 = "^2.9.11"
itsdangerous = "^2.2.0"
fastapi-mail = "^1.5.8"
pyjwt = "^2.10.1"
argon2-cffi = "^25.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
